synthetica/requirements.txt
# Synthetica — Crisis Management Swarm (Google Cloud / Gemini)
# Python 3.9+

redis>=5.0.0
pydantic>=2.0.0
google-generativeai>=0.8.0
fakeredis>=2.0.0
google-generativeai
google-cloud-aiplatform
requests>=2.0.0
